# docs/swagger.yaml
basePath: /
definitions:
...
  gos_internal_application_usecase.SyncExecutorParamDefsOutput:
    properties:
      created:
        type: integer
      inactivated:
        type: integer
      skipped:
        type: integer
      total:
        type: integer
      updated:
        type: integer
    type: object
...
  internal_interfaces_http.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  internal_interfaces_http.ExecutorParamDefDataResponse:
    properties:
      data:
        $ref: '#/definitions/internal_interfaces_http.ExecutorParamDefResponse'
    type: object
  internal_interfaces_http.ExecutorParamDefListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_interfaces_http.ExecutorParamDefResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  internal_interfaces_http.ExecutorParamDefResponse:
    properties:
      application_id:
        type: string
      application_key:
        type: string
      application_name:
        type: string
      binding_type:
        type: string
      can_edit:
        type: boolean
      can_view:
        type: boolean
      created_at:
        type: string
      default_value:
        type: string
      description:
        type: string
      editable:
        type: boolean
      executor_param_name:
        type: string
      executor_type:
        type: string
      id:
        type: string
      param_key:
        type: string
      param_type:
        type: string
      pipeline_id:
        type: string
      pipeline_name:
        type: string
      raw_meta:
        type: string
      required:
        type: boolean
      single_select:
        type: boolean
      sort_no:
        type: integer
      source_from:
        type: string
      status:
        type: string
      updated_at:
        type: string
      visible:
        type: boolean
    type: object
...
  internal_interfaces_http.SyncExecutorParamDefsResponse:
    properties:
      data:
        $ref: '#/definitions/gos_internal_application_usecase.SyncExecutorParamDefsOutput'
    type: object
...
  internal_interfaces_http.UpdateExecutorParamDefRequest:
    properties:
      param_key:
        type: string
    type: object
...
info:
  contact: {}
  description: Internal deployment platform API.
  title: GOS Release API
  version: "1.0"
paths:
...
  /applications/{id}/executor-param-defs:
    get:
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Binding type, default ci
        in: query
        name: binding_type
        type: string
      - description: Visible flag
        in: query
        name: visible
        type: boolean
      - description: Editable flag
        in: query
        name: editable
        type: boolean
      - description: Mapped platform param key
        in: query
        name: param_key
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.ExecutorParamDefListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: List bound executor param definitions by application
      tags:
      - executor-param-defs
...
  /executor-param-defs:
    get:
      parameters:
      - description: Application ID
        in: query
        name: application_id
        type: string
      - description: Application name, application key, or platform key
        in: query
        name: keyword
        type: string
      - description: Binding type, default ci
        in: query
        name: binding_type
        type: string
      - description: Visible flag
        in: query
        name: visible
        type: boolean
      - description: Editable flag
        in: query
        name: editable
        type: boolean
      - description: Status
        in: query
        name: status
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.ExecutorParamDefListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: List executor param definitions across applications
      tags:
      - executor-param-defs
  /executor-param-defs/{id}:
    get:
      parameters:
      - description: Executor param definition ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.ExecutorParamDefDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: Get executor param definition by ID
      tags:
      - executor-param-defs
    put:
      consumes:
      - application/json
      parameters:
      - description: Executor param definition ID
        in: path
        name: id
        required: true
        type: string
      - description: Update executor param definition request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http.UpdateExecutorParamDefRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.ExecutorParamDefDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: Update executor param definition mapping
      tags:
      - executor-param-defs
...
  /jenkins/executor-param-defs/sync:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.SyncExecutorParamDefsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: Sync executor param definitions from Jenkins
      tags:
      - executor-param-defs
...
  /pipelines/{id}/param-defs:
    get:
      parameters:
      - description: Pipeline ID
        in: path
        name: id
        required: true
        type: string
      - description: Executor type
        in: query
        name: executor_type
        type: string
      - description: Visible flag
        in: query
        name: visible
        type: boolean
      - description: Editable flag
        in: query
        name: editable
        type: boolean
      - description: Mapped platform param key
        in: query
        name: param_key
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.ExecutorParamDefListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: List executor param definitions
      tags:
      - executor-param-defs
...
schemes:
- http
- https
swagger: "2.0"